어댑터 리스트어댑터로 변경
