+++ app/src/main/java/com/android/sabsigan/main/chatting/MessageAdapter.kt
@@ -4,20 +4,21 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.databinding.ViewDataBinding
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.android.sabsigan.data.ChatMessage
+import com.android.sabsigan.data.User
 import com.android.sabsigan.databinding.AdapterMyMessageBinding
 import com.android.sabsigan.databinding.AdapterOtherMessageBinding
 import com.android.sabsigan.viewModel.ChatViewModel
 
 
-class MessageAdapter(private val viewModel: ChatViewModel): RecyclerView.Adapter<MessageAdapter.MessageViewHolder>() {
-    var messageList = arrayListOf<ChatMessage>()
-
+class MessageAdapter(private val viewModel: ChatViewModel): ListAdapter<ChatMessage, MessageAdapter.MessageViewHolder>(diffUtil) {
     val TYPE_MY = 0
     val TYPE_OTHER = 1
 
-    class MessageViewHolder private constructor(var binding: ViewDataBinding, val viewType: Int) : RecyclerView.ViewHolder(binding.root) {
+    class MessageViewHolder(var binding: ViewDataBinding, val viewType: Int) : RecyclerView.ViewHolder(binding.root) {
         fun bind(viewModel: ChatViewModel, chatMessage: ChatMessage) {
             if (viewType == 0) {
                 (binding as AdapterMyMessageBinding).chatMessage = chatMessage
@@ -29,20 +30,10 @@ class MessageAdapter(private val viewModel: ChatViewModel): RecyclerView.Adapter
                 binding.executePendingBindings()
             }
         }
-
-        companion object {
-            fun from(parent: ViewGroup, viewType: Int) : MessageViewHolder {
-                val layoutInflater = LayoutInflater.from(parent.context)
-                val binding = if (viewType == 0) AdapterMyMessageBinding.inflate(layoutInflater, parent, false)
-                                else  AdapterOtherMessageBinding.inflate(layoutInflater, parent, false)
-
-                return MessageViewHolder(binding, viewType)
-            }
-        }
     }
 
     override fun getItemViewType(position: Int): Int {
-        if(messageList[position].uid.equals(viewModel.getUID())) {
+        if(getItem(position).uid.equals(viewModel.getUID())) {
             Log.d("ViewType", "0")
             return TYPE_MY;
         } else {
@@ -52,12 +43,24 @@ class MessageAdapter(private val viewModel: ChatViewModel): RecyclerView.Adapter
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MessageViewHolder {
-        return MessageViewHolder.from(parent, viewType)
+        val layoutInflater = LayoutInflater.from(parent.context)
+        val binding = if (viewType == 0) AdapterMyMessageBinding.inflate(layoutInflater, parent, false)
+        else  AdapterOtherMessageBinding.inflate(layoutInflater, parent, false)
+
+        return MessageViewHolder(binding, viewType)
     }
 
     override fun onBindViewHolder(holder: MessageViewHolder, position: Int) {
-        holder.bind(viewModel, messageList[position])
+        holder.bind(viewModel, getItem(position))
     }
 
-    override fun getItemCount() = messageList.size
+    companion object {
+        val diffUtil = object : DiffUtil.ItemCallback<ChatMessage>() {
+            override fun areItemsTheSame(oldItem: ChatMessage, newItem: ChatMessage) =
+                oldItem.id == newItem.id
+
+            override fun areContentsTheSame(oldItem: ChatMessage, newItem: ChatMessage) =
+                oldItem == newItem
+        }
+    }
 }
