+++ app/src/main/java/com/android/sabsigan/main/user/UserListBindingAdapter.kt
@@ -4,21 +4,14 @@ import android.annotation.SuppressLint
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.android.sabsigan.data.User
+import okhttp3.internal.notify
 
 object UserListBindingAdapter {
     @BindingAdapter("app:userItems")
     @JvmStatic
-    fun setItems(recyclerView: RecyclerView, items: ArrayList<User>?) {
-//        if(recyclerView.adapter == null) {
-//            val adapter = UserListAdapter()
-//            adapter.setHasStableIds(true)
-//            recyclerView.adapter = adapter
-//        }
-
+    fun setItems(recyclerView: RecyclerView, items: List<User>?) {
         items?.let {
-            val adapter = recyclerView.adapter as UserListAdapter
-            adapter.userList = items
-            adapter.notifyDataSetChanged()
+            (recyclerView.adapter as UserListAdapter).submitList(it)
         }
     }
 }
