+++ app/src/main/res/layout/activity_chat.xml
@@ -41,7 +41,6 @@
             android:layout_width="match_parent"
             android:layout_height="0dp"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-            app:stackFromEnd="true"
             app:layout_constraintTop_toBottomOf="@+id/toolbar"
             app:layout_constraintBottom_toTopOf="@+id/inputLayout"
             app:messageItems="@{viewModel.messageList}"/>
