+++ app/src/main/res/layout/adapter_user_list.xml
@@ -2,7 +2,9 @@
 <layout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".main.user.UserListAdapter">
+
     <data>
         <variable
             name="user"
