+++ app/src/main/res/layout/fragment_user.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".main.user.UserFragment">
 
     <data>
         <variable
@@ -11,8 +12,7 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:context=".main.user.UserFragment">
+        android:layout_height="match_parent">
 
         <TextView
             android:id="@+id/text_user"
@@ -48,8 +48,7 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            android:onClick="@{()->viewModel.clickKKK()}"
-            />
+            android:onClick="@{()->viewModel.clickKKK()}"/>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
